refactor: TimeTableScreen uiState 구성 + 신규 ViewModel 추가 (#513)

+++ app/src/main/java/com/wafflestudio/snutt2/domainmodel/LectureColor.kt
@@ -24,6 +24,6 @@ sealed interface LectureColor {
     }
 
     data class BuiltIn(
-        val colorIndex: Long,
+        val colorIndex: Int,
     ) : LectureColor
 }
